+++ aiohomematic/model/calculated/operating_voltage_level.py
@@ -10,7 +10,7 @@ from enum import StrEnum
 import logging
 from typing import Any, Final
 
-from aiohomematic.const import CalulatedParameter, DataPointCategory, Parameter, ParameterType, ParamsetKey
+from aiohomematic.const import CalculatedParameter, DataPointCategory, Parameter, ParameterType, ParamsetKey
 from aiohomematic.interfaces.model import ChannelProtocol
 from aiohomematic.model.calculated.data_point import CalculatedDataPoint
 from aiohomematic.model.calculated.support import calculate_operating_voltage_level
@@ -38,7 +38,7 @@ class OperatingVoltageLevel[SensorT: float | None](CalculatedDataPoint[SensorT])
         "_voltage_max",
     )
 
-    _calculated_parameter = CalulatedParameter.OPERATING_VOLTAGE_LEVEL
+    _calculated_parameter = CalculatedParameter.OPERATING_VOLTAGE_LEVEL
     _category = DataPointCategory.SENSOR
 
     def __init__(self, *, channel: ChannelProtocol) -> None: